Move Mastic examples to their own file and clean up imports
