+++ poc/mastic.py
@@ -5,7 +5,7 @@ from typing import Optional, Sequence, TypeAlias, TypeVar, cast
 from vdaf_poc.common import (byte, concat, front, to_be_bytes, vec_add,
                              vec_sub, zeros)
 from vdaf_poc.field import NttField
-from vdaf_poc.flp_bbcggi19 import Count, FlpBBCGGI19, Sum, Valid
+from vdaf_poc.flp_bbcggi19 import FlpBBCGGI19, Valid
 from vdaf_poc.vdaf import Vdaf
 from vdaf_poc.xof import XofTurboShake128
 
@@ -489,372 +489,3 @@ class Mastic(
             to_be_bytes(self.ID, 4),
             to_be_bytes(usage, 2),
         ])
-
-
-def get_reports_from_measurements(mastic, measurements):
-    from vdaf_poc.common import gen_rand
-
-    reports = []
-    for measurement in measurements:
-        nonce = gen_rand(16)
-        rand = gen_rand(mastic.RAND_SIZE)
-        (public_share, input_shares) = mastic.shard(measurement, nonce, rand)
-        reports.append((nonce, public_share, input_shares))
-    return reports
-
-
-def get_threshold(thresholds, prefix, level):
-    '''
-    Return the threshold of the given (prefix, level) if the tuple exists. If
-    not, check if any of its prefixes exist. If not, return the default
-    threshold.
-    '''
-    while level > 0:
-        if (prefix, level) in thresholds:
-            return thresholds[(prefix, level)]
-        prefix >>= 1
-        level -= 1
-    return thresholds['default']  # Return the default threshold
-
-
-def compute_heavy_hitters(mastic, bits, thresholds, reports):
-    from vdaf_poc.common import gen_rand
-
-    verify_key = gen_rand(16)
-
-    prefixes = [0, 1]
-    prev_agg_params = []
-    heavy_hitters = []
-    for level in range(bits):
-        agg_param = (level, prefixes, level == 0)
-        assert mastic.is_valid(agg_param, prev_agg_params)
-
-        # Aggregators prepare reports for aggregation.
-        out_shares = [[], []]
-        for (nonce, public_share, input_shares) in reports:
-            # Each aggregator broadcast its prep share; ...
-            (prep_state, prep_shares) = zip(*[
-                mastic.prep_init(
-                    verify_key,
-                    agg_id,
-                    agg_param,
-                    nonce,
-                    public_share,
-                    input_shares[agg_id]) for agg_id in [0, 1]
-            ])
-
-            # computes the prep message; ...
-            prep_msg = mastic.prep_shares_to_prep(agg_param, prep_shares)
-
-            # and computes its output share.
-            for agg_id in [0, 1]:
-                out_shares[agg_id].append(
-                    mastic.prep_next(prep_state[agg_id], prep_msg))
-
-        # Aggregators aggregate their output shares.
-        agg_shares = [
-            mastic.aggregate(agg_param, out_shares[agg_id]) for agg_id in [0, 1]
-        ]
-
-        # Collector computes the aggregate result.
-        agg_result = mastic.unshard(agg_param, agg_shares, len(reports))
-        prev_agg_params.append(agg_param)
-
-        if level < bits - 1:
-            # Compute the next set of candidate prefixes.
-            next_prefixes = []
-            for (prefix, count) in zip(prefixes, agg_result):
-                threshold = get_threshold(thresholds, prefix, level)
-                if count >= threshold:
-                    next_prefixes.append(prefix << 1)
-                    next_prefixes.append((prefix << 1) | 1)
-            prefixes = next_prefixes
-        else:
-            for (prefix, count) in zip(prefixes, agg_result):
-                threshold = get_threshold(thresholds, prefix, level)
-                if count >= threshold:
-                    heavy_hitters.append(prefix)
-    return heavy_hitters
-
-
-def example_weighted_heavy_hitters_mode():
-    from vdaf_poc.field import Field64
-
-    bits = 4
-    mastic = Mastic(bits, Count(Field64))
-
-    # Clients shard their measurements. Each measurement is comprised of
-    # `(alpha, beta)` where `alpha` is the payload string and `beta` is its
-    # weight. Here the weight is simply a counter (either `0` or `1`).
-    measurements = [
-        (0b1001, 1),
-        (0b0000, 1),
-        (0b0000, 0),
-        (0b0000, 1),
-        (0b1001, 1),
-        (0b0000, 1),
-        (0b1100, 1),
-        (0b0011, 1),
-        (0b1111, 0),
-        (0b1111, 0),
-        (0b1111, 1),
-    ]
-
-    reports = get_reports_from_measurements(mastic, measurements)
-
-    thresholds = {
-        'default': 2,
-    }
-
-    # Collector and Aggregators compute the weighted heavy hitters.
-    heavy_hitters = compute_heavy_hitters(mastic, bits, thresholds, reports)
-    print("Weighted heavy-hitters:",
-          list(map(lambda x: bin(x), heavy_hitters)))
-    assert heavy_hitters == [0, 9]
-
-
-def example_weighted_heavy_hitters_mode_with_different_thresholds():
-    from vdaf_poc.field import Field64
-
-    bits = 4
-    mastic = Mastic(bits, Count(Field64))
-
-    # Clients shard their measurements. Each measurement is comprised of
-    # `(alpha, beta)` where `alpha` is the payload string and `beta` is its
-    # weight. Here the weight is simply a counter (either `0` or `1`).
-    measurements = [
-        (0b0000, 1),
-        (0b0001, 1),
-        (0b1001, 1),
-        (0b1001, 1),
-        (0b1010, 1),
-        (0b1010, 1),
-        (0b1111, 1),
-        (0b1111, 1),
-        (0b1111, 1),
-        (0b1111, 1),
-        (0b1111, 1),
-    ]
-
-    reports = get_reports_from_measurements(mastic, measurements)
-
-    # (prefix, level): threshold
-    # Note that levels start from zero
-    thresholds = {
-        'default': 2,
-        (0b00, 1): 1,
-        (0b10, 1): 3,
-        (0b11, 1): 5,
-    }
-
-    # Collector and Aggregators compute the weighted heavy hitters.
-    heavy_hitters = compute_heavy_hitters(mastic, bits, thresholds, reports)
-    print("Weighted heavy-hitters with different thresholds:",
-          list(map(lambda x: bin(x), heavy_hitters)))
-    assert heavy_hitters == [0, 1, 15]
-
-
-def example_attribute_based_metrics_mode():
-    import hashlib
-
-    from vdaf_poc.common import gen_rand
-    from vdaf_poc.field import Field64
-
-    bits = 8
-    mastic = Mastic(bits, Sum(Field64, 3))
-    verify_key = gen_rand(16)
-
-    def h(attr):
-        """
-        Hash the attribute to a fixed-size string whose size matches the
-        bit-size for our instance of Mastic. For testing purposes, we truncate
-        to the first `8` bits of the hash; in practice we would need collision
-        resistance. Mastic should be reasonably fast even for `bits == 256`
-        (the same as SHA-3).
-        """
-        assert bits == 8
-        sha3 = hashlib.sha3_256()
-        sha3.update(attr.encode('ascii'))
-        return sha3.digest()[0]
-
-    # Clients shard their measurements. Each measurement is comprised of
-    # (`alpha`, `beta`) where `beta` is the Client's contribution to the
-    # aggregate with attribute `alpha`.
-    #
-    # In this example, each Client casts a "vote" (between '0' and '3') and
-    # attributes their vote with their home country.
-    measurements = [
-        ('United States', 1),
-        ('Greece', 1),
-        ('United States', 2),
-        ('Greece', 0),
-        ('United States', 0),
-        ('India', 1),
-        ('Greece', 0),
-        ('United States', 1),
-        ('Greece', 1),
-        ('Greece', 3),
-        ('Greece', 1),
-    ]
-    reports = []
-    for (attr, vote) in measurements:
-        nonce = gen_rand(16)
-        rand = gen_rand(mastic.RAND_SIZE)
-        (public_share, input_shares) = mastic.shard(
-            (h(attr), vote),
-            nonce,
-            rand,
-        )
-        reports.append((nonce, public_share, input_shares))
-
-    # Aggregators aggregate the reports, breaking them down by home country.
-    attrs = [
-        'Greece',
-        'Mexico',
-        'United States',
-    ]
-    agg_param = (bits-1, list(map(lambda attr: h(attr), attrs)), True)
-    assert mastic.is_valid(agg_param, [])
-
-    # Aggregators prepare reports for aggregation.
-    out_shares = [[], []]
-    for (nonce, public_share, input_shares) in reports:
-        # Each aggregator broadcast its prep share; ...
-        (prep_state, prep_shares) = zip(*[
-            mastic.prep_init(
-                verify_key,
-                agg_id,
-                agg_param,
-                nonce,
-                public_share,
-                input_shares[agg_id]) for agg_id in [0, 1]
-        ])
-
-        # computes the prep message; ...
-        prep_msg = mastic.prep_shares_to_prep(agg_param, prep_shares)
-
-        # and computes its output share.
-        for agg_id in [0, 1]:
-            out_shares[agg_id].append(
-                mastic.prep_next(prep_state[agg_id], prep_msg))
-
-    # Aggregators aggregate their output shares.
-    agg_shares = [
-        mastic.aggregate(agg_param, out_shares[agg_id]) for agg_id in [0, 1]
-    ]
-
-    # Collector computes the aggregate result.
-    agg_result = mastic.unshard(agg_param, agg_shares, len(measurements))
-    print('Election results:', list(zip(attrs, agg_result)))
-    assert agg_result == [6, 0, 4]
-
-
-def example_poplar1_overhead():
-    import math
-
-    from vdaf_poc.common import gen_rand
-    from vdaf_poc.field import Field64, Field128
-    from vdaf_poc.flp_bbcggi19 import Histogram
-    from vdaf_poc.vdaf_poplar1 import Poplar1
-    from vdaf_poc.vdaf_prio3 import Prio3Histogram
-
-    nonce = gen_rand(16)
-
-    cls = Poplar1(256)
-    (public_share, input_shares) = cls.shard(0, nonce, gen_rand(cls.RAND_SIZE))
-    b = 0
-    p = len(cls.test_vec_encode_public_share(public_share))
-    b += p
-    print('Poplar1(256) public share len:', p)
-    p = len(cls.test_vec_encode_input_share(input_shares[0]))
-    b += p
-    print('Poplar1(256) input share 0 len:', p)
-    p = len(cls.test_vec_encode_input_share(input_shares[1]))
-    b += p
-    print('Poplar1(256) input share 1 len:', p)
-    poplar1_bytes_uploaded = b
-
-    cls = Mastic(256, Count(Field64))
-    (public_share, input_shares) = cls.shard((0, 0),
-                                             nonce,
-                                             gen_rand(cls.RAND_SIZE))
-    b = 0
-    p = len(cls.test_vec_encode_public_share(public_share))
-    b += p
-    print('Mastic(256,Count()) public share len:', p)
-    p = len(cls.test_vec_encode_input_share(input_shares[0]))
-    b += p
-    print('Mastic(256,Count()) input share 0 len:', p)
-    p = len(cls.test_vec_encode_input_share(input_shares[1]))
-    b += p
-    print('Mastic(256,Count()) input share 1 len:', p)
-    mastic_count_bytes_uploaded = b
-
-    cls = Mastic(256, Sum(Field64, 8))
-    (public_share, input_shares) = cls.shard((0, 0),
-                                             nonce,
-                                             gen_rand(cls.RAND_SIZE))
-    b = 0
-    p = len(cls.test_vec_encode_public_share(public_share))
-    b += p
-    print('Mastic(256,Sum(8)) public share len:', p)
-    p = len(cls.test_vec_encode_input_share(input_shares[0]))
-    b += p
-    print('Mastic(256,Sum(8)) input share 0 len:', p)
-    p = len(cls.test_vec_encode_input_share(input_shares[1]))
-    b += p
-    print('Mastic(256,Sum(8)) input share 1 len:', p)
-    mastic_sum8_bytes_uploaded = b
-
-    print('Mastic(256,Count()) overhead for Poplar1(256): {:.2f}%'.format(
-        mastic_count_bytes_uploaded / poplar1_bytes_uploaded * 100))
-    print('Mastic(256,Sum(8)) overhead for Mastic(256,Count()): {:.2f}%'.format(
-        mastic_sum8_bytes_uploaded / mastic_count_bytes_uploaded * 100))
-
-    cls = Mastic(32, Histogram(Field128, 100, 10))
-    (public_share, input_shares) = cls.shard((0, 0),
-                                             nonce,
-                                             gen_rand(cls.RAND_SIZE))
-    b = 0
-    p = len(cls.test_vec_encode_public_share(public_share))
-    b += p
-    print('Mastic(32,Histogram(100, 10)) public share len:', p)
-    p = len(cls.test_vec_encode_input_share(input_shares[0]))
-    b += p
-    print('Mastic(32,Histogram(100, 10)) input share 0 len:', p)
-    p = len(cls.test_vec_encode_input_share(input_shares[1]))
-    b += p
-    print('Mastic(32,Histogram(100, 10)) input share 1 len:', p)
-    print('Mastic(32,Histogram(100, 10)) total upload len:', b)
-    mastic_hist_bytes_uploaded = b
-
-    length = 100 * 100  # base histogram length * number of attributes
-    chunk_length = math.floor(math.sqrt(length))
-    cls = Prio3Histogram(2, length, chunk_length)
-    (public_share, input_shares) = cls.shard(0, nonce, gen_rand(cls.RAND_SIZE))
-    b = 0
-    p = len(cls.test_vec_encode_public_share(public_share))
-    b += p
-    print('Prio3Histogram({}, {}) public share len:'.format(
-        length, chunk_length), p)
-    p = len(cls.test_vec_encode_input_share(input_shares[0]))
-    b += p
-    print('Prio3Histogram({}, {}) input share 0 len:'.format(
-        length, chunk_length), p)
-    p = len(cls.test_vec_encode_input_share(input_shares[1]))
-    b += p
-    print('Prio3Histogram({}, {}) input share 1 len:'.format(
-        length, chunk_length), p)
-    print('Prio3Histogram({}, {}) total upload len:'.format(
-        length, chunk_length), b)
-    prio3_hist_bytes_uploaded = b
-
-    print(prio3_hist_bytes_uploaded / mastic_hist_bytes_uploaded)
-
-
-# TODO Move this to a separate file.
-if __name__ == '__main__':
-    example_poplar1_overhead()
-    example_weighted_heavy_hitters_mode()
-    example_attribute_based_metrics_mode()
-    example_weighted_heavy_hitters_mode_with_different_thresholds()
